feat: separate CompanyData and FilingData schemas from SEC EDGAR integration
